requirements.txt
python-gyazo
opencv-python
pyautogui
webbrowser